+++ hammond-gtk/src/widgets/show.rs
@@ -1,5 +1,6 @@
 use dissolve;
 use failure::Error;
+use glib;
 use gtk;
 use gtk::prelude::*;
 use open;
@@ -12,6 +13,8 @@ use app::Action;
 use utils::get_pixbuf_from_path;
 use widgets::episode::episodes_listbox;
 
+use std::cell::RefCell;
+use std::rc::Rc;
 use std::sync::Arc;
 use std::sync::mpsc::Sender;
 use std::thread;
@@ -26,6 +29,9 @@ pub struct ShowWidget {
     settings: gtk::MenuButton,
     unsub: gtk::Button,
     episodes: gtk::Frame,
+    notif: gtk::Revealer,
+    notif_label: gtk::Label,
+    notif_undo: gtk::Button,
 }
 
 impl Default for ShowWidget {
@@ -41,6 +47,10 @@ impl Default for ShowWidget {
         let link: gtk::Button = builder.get_object("link_button").unwrap();
         let settings: gtk::MenuButton = builder.get_object("settings_button").unwrap();
 
+        let notif: gtk::Revealer = builder.get_object("notif_revealer").unwrap();
+        let notif_label: gtk::Label = builder.get_object("notif_label").unwrap();
+        let notif_undo: gtk::Button = builder.get_object("undo_button").unwrap();
+
         ShowWidget {
             container,
             scrolled_window,
@@ -50,6 +60,9 @@ impl Default for ShowWidget {
             link,
             settings,
             episodes,
+            notif,
+            notif_label,
+            notif_undo,
         }
     }
 }
@@ -62,6 +75,8 @@ impl ShowWidget {
     }
 
     pub fn init(&self, pd: Arc<Podcast>, sender: Sender<Action>) {
+        let builder = gtk::Builder::new_from_resource("/org/gnome/hammond/gtk/show_widget.ui");
+
         // Hacky workaround so the pd.id() can be retrieved from the `ShowStack`.
         WidgetExt::set_name(&self.container, &pd.id().to_string());
 
@@ -88,6 +103,25 @@ impl ShowWidget {
                 error!("Error: {}", err);
             }
         });
+
+        let show_menu: gtk::Popover = builder.get_object("show_menu").unwrap();
+        let mark_all: gtk::ModelButton = builder.get_object("mark_all_watched").unwrap();
+
+        let notif = self.notif.clone();
+        let notif_label = self.notif_label.clone();
+        let notif_undo = self.notif_undo.clone();
+        let episodes = self.episodes.clone();
+        mark_all.connect_clicked(clone!(pd, sender => move |_| {
+            on_played_button_clicked(
+                pd.clone(),
+                &notif,
+                &notif_label,
+                &notif_undo,
+                &episodes,
+                sender.clone()
+            )
+        }));
+        self.settings.set_popover(&show_menu);
     }
 
     /// Populate the listbox with the shows episodes.
@@ -142,9 +176,83 @@ fn on_unsub_button_clicked(
     Ok(())
 }
 
-#[allow(dead_code)]
-fn on_played_button_clicked(pd: &Podcast, sender: Sender<Action>) -> Result<(), Error> {
+fn on_played_button_clicked(
+    pd: Arc<Podcast>,
+    notif: &gtk::Revealer,
+    label: &gtk::Label,
+    undo: &gtk::Button,
+    episodes: &gtk::Frame,
+    sender: Sender<Action>,
+) {
+    if dim_titles(episodes).is_none() {
+        error!("Something went horribly wrong when dimming the titles.");
+        warn!("RUN WHILE YOU STILL CAN!");
+    }
+
+    label.set_text("All episodes where marked as watched.");
+    notif.set_reveal_child(true);
+
+    // Set up the callback
+    let id = timeout_add_seconds(
+        10,
+        clone!(sender => move || {
+        if let Err(err) = wrap(&pd, sender.clone()) {
+            error!(
+                "Something went horribly wrong with the notif callback: {}",
+                err
+            );
+        }
+        glib::Continue(false)
+    }),
+    );
+
+    let id = Rc::new(RefCell::new(Some(id)));
+
+    // Cancel the callback
+    undo.connect_clicked(clone!(id, notif, sender => move |_| {
+        let foo = id.borrow_mut().take();
+        if let Some(id) = foo {
+            glib::source::source_remove(id);
+            notif.set_reveal_child(false);
+            if let Err(err) = sender.send(Action::RefreshWidgetIfVis) {
+                error!("Something went horribly wrong with the Action channel: {}", err)
+            }
+        }
+    }));
+}
+
+fn wrap(pd: &Podcast, sender: Sender<Action>) -> Result<(), Error> {
     dbqueries::update_none_to_played_now(pd)?;
-    sender.send(Action::RefreshWidget)?;
+    sender.send(Action::RefreshWidgetIfVis)?;
+    sender.send(Action::RefreshEpisodesView)?;
     Ok(())
+}
+
+// Ideally if we had a custom widget this would have been as simple as:
+// `for row in listbox { ep = row.get_episode(); ep.dim_title(); }`
+// But now I can't think of a better way to do it than hardcoding the title
+// position relative to the EpisodeWidget container gtk::Box.
+fn dim_titles(episodes: &gtk::Frame) -> Option<()> {
+    let listbox = episodes
+        .get_children()
+        .remove(0)
+        .downcast::<gtk::ListBox>()
+        .ok()?;
+    let children = listbox.get_children();
+
+    for row in children {
+        let row = row.downcast::<gtk::ListBoxRow>().ok()?;
+        let container = row.get_children().remove(0).downcast::<gtk::Box>().ok()?;
+        let foo = container
+            .get_children()
+            .remove(0)
+            .downcast::<gtk::Box>()
+            .ok()?;
+        let bar = foo.get_children().remove(0).downcast::<gtk::Box>().ok()?;
+        let baz = bar.get_children().remove(0).downcast::<gtk::Box>().ok()?;
+        let title = baz.get_children().remove(0).downcast::<gtk::Label>().ok()?;
+
+        title.get_style_context().map(|c| c.add_class("dim-label"));
+    }
+    Some(())
 }